feat: Update id_procedimento_unificado to id_procedimento_interno in sisreg schema
This commit updates the sisreg schema in the raw/sisreg model to change the name of the "id_procedimento_unificado" field to "id_procedimento_interno". The description of the field is also updated to reflect this change. Additionally, a new field "id_procedimento_unificado" is added with a description indicating it as an alternative unique identifier for the procedure being performed.