refactor: Update database and table names in _sih_sources.yml and raw_sih__indicadores_hospitalares.sql
This commit updates the database and table names in the _sih_sources.yml and raw_sih__indicadores_hospitalares.sql files. The database name is changed from "rj-sms-dev" to "rj-sms" and the table name is changed from "indicadores_hospitalares_parquet" to "indicadores_hospitalares". This change aligns the names with the current configuration and improves consistency in the codebase.