feat: Add data_particao field to _mart_estoque_schema.yml
This commit adds the data_particao field to the _mart_estoque_schema.yml file. The data_particao field indicates the date of reference for the position. This field is quoted and has a data_type of date. Additionally, the description for the material_quantidade field has been updated to include information about the reference date (data_particao).