Update historico_clinico_schema.yml and datasus_schema.yml
This commit updates the historico_clinico_schema.yml and datasus_schema.yml files. It adds policy tags to various fields in the historico_clinico_schema.yml file, indicating the data sensitivity and classification. Additionally, it updates the description of the raw_datasus__cid10 table in the datasus_schema.yml file to clarify that it contains the Classificação Internacional de Doenças (CID-10).